Consistently use the default unit (e.g bitcoin) in the database instead of some values being stored as satoshis and some as the default unit.

+++ blockchain-analytics-flink/src/main/java/io/github/anttikaikkonen/blockchainanalyticsflink/casssandra/models/AddressBalance.java
@@ -22,6 +22,6 @@ public class AddressBalance {
     @ClusteringColumn(0)
     private Date timestamp;
     
-    private long balance;
+    private double balance;
     
 }
